requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn>=0.22.0
apscheduler>=3.10.0
aiohttp>=3.8.0

# Optional: for live Polymarket orders
# py_clob_client>=0.1.0

# Optional: for live Kalshi orders (RSA signing)
# cryptography>=41.0.0

# Optional: for LLM validation (CrewAI/LangChain)
# crewai>=0.1.0
# langchain-openai>=0.0.5
